@@ -1,126 +1,88 @@
 {
   "name": "llama-wool-farm",
   "version": "1.0.0",
-  "description": "A progressive web app idle clicker game about llamas producing wool",
-  "main": "src/index.ts",
+  "description": "A comprehensive llama wool production management system",
+  "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "webpack serve --mode development --config webpack.config.js",
-    "dev:https": "npm run generate-ssl && webpack serve --mode development --config webpack.config.js",
-    "build": "webpack --mode production --config webpack.config.js",
-    "build:analyze": "webpack --mode production --config webpack.config.js --analyze",
-    "build:dev": "webpack --mode development --config webpack.config.js",
-    "test": "jest",
-    "test:unit": "jest --testPathPattern=tests/unit",
-    "test:integration": "jest --testPathPattern=tests/integration",
-    "test:e2e": "cypress run",
-    "test:coverage": "jest --coverage",
-    "lint": "eslint src --ext .ts,.tsx",
-    "lint:fix": "eslint src --ext .ts,.tsx --fix",
-    "type-check": "tsc --noEmit",
-    "generate-icons": "node scripts/generate-icons.js",
-    "generate-ssl": "node scripts/generate-ssl.js",
-    "serve": "serve -s dist -p 8080",
-    "serve:https": "serve -s dist -p 8080 --ssl-cert ssl/server.crt --ssl-key ssl/server.key",
-    "clean": "rimraf dist coverage .webpack-cache",
-    "validate": "npm run lint && npm run type-check && npm run test",
-    "prepare": "husky install"
-  },
-  "keywords": [
-    "game",
-    "phaser",
-    "pwa",
-    "idle",
-    "clicker",
-    "typescript"
-  ],
-  "author": "Llama Wool Farm Team",
-  "license": "MIT",
-  "dependencies": {
-    "@reduxjs/toolkit": "^2.0.1",
-    "axios": "^1.6.5",
-    "decimal.js": "^10.4.3",
-    "localforage": "^1.10.0",
-    "phaser": "^3.80.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^9.0.4",
-    "socket.io-client": "^4.7.4",
-    "workbox-background-sync": "^7.0.0",
-    "workbox-cacheable-response": "^7.0.0",
-    "workbox-core": "^7.0.0",
-    "workbox-expiration": "^7.0.0",
-    "workbox-navigation-preload": "^7.0.0",
-    "workbox-precaching": "^7.0.0",
-    "workbox-routing": "^7.0.0",
-    "workbox-strategies": "^7.0.0",
-    "workbox-webpack-plugin": "^7.0.0"
+    "dev": "concurrently \"npm run dev:frontend\" \"npm run dev:backend\"",
+    "dev:frontend": "cd frontend && npm run dev",
+    "dev:backend": "cd backend && npm run dev",
+    "build": "npm run build:frontend && npm run build:backend",
+    "build:frontend": "cd frontend && npm run build",
+    "build:backend": "cd backend && npm run build",
+    "test": "jest --coverage",
+    "test:watch": "jest --watch",
+    "test:frontend": "cd frontend && npm test",
+    "test:backend": "cd backend && npm test",
+    "lint": "eslint . --ext .ts,.tsx,.js,.jsx",
+    "lint:fix": "eslint . --ext .ts,.tsx,.js,.jsx --fix",
+    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,css,scss,md}\"",
+    "format:check": "prettier --check \"**/*.{js,jsx,ts,tsx,json,css,scss,md}\"",
+    "typecheck": "tsc --noEmit",
+    "prepare": "husky install",
+    "install:frontend": "cd frontend && npm install",
+    "install:backend": "cd backend && npm install",
+    "clean": "rimraf node_modules frontend/node_modules backend/node_modules",
+    "clean:build": "rimraf frontend/dist backend/dist"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.7",
-    "@babel/preset-env": "^7.23.8",
-    "@babel/preset-react": "^7.23.3",
-    "@babel/preset-typescript": "^7.23.3",
-    "@testing-library/jest-dom": "^6.2.0",
-    "@testing-library/react": "^14.1.2",
-    "@types/jest": "^29.5.11",
-    "@types/node": "^20.11.5",
-    "@types/react": "^18.2.48",
-    "@types/react-dom": "^18.2.18",
-    "@typescript-eslint/eslint-plugin": "^6.19.0",
-    "@typescript-eslint/parser": "^6.19.0",
-    "autoprefixer": "^10.4.16",
-    "babel-loader": "^9.1.3",
-    "canvas": "^2.11.2",
-    "clean-webpack-plugin": "^4.0.0",
-    "compression-webpack-plugin": "^11.0.0",
-    "copy-webpack-plugin": "^12.0.2",
-    "css-loader": "^6.9.1",
-    "css-minimizer-webpack-plugin": "^5.0.1",
-    "cssnano": "^6.0.3",
-    "cypress": "^13.6.3",
-    "eslint": "^8.56.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.24",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "concurrently": "^8.2.2",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^27.9.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "html-webpack-plugin": "^5.6.0",
-    "husky": "^8.0.3",
+    "husky": "^9.0.11",
     "jest": "^29.7.0",
-    "jest-canvas-mock": "^2.5.2",
-    "lint-staged": "^15.2.0",
-    "mini-css-extract-plugin": "^2.7.7",
-    "postcss": "^8.4.33",
-    "postcss-loader": "^7.3.4",
-    "prettier": "^3.2.4",
+    "lint-staged": "^15.2.2",
+    "prettier": "^3.2.5",
     "rimraf": "^5.0.5",
-    "serve": "^14.2.1",
-    "sharp": "^0.33.5",
-    "style-loader": "^3.3.4",
-    "terser-webpack-plugin": "^5.3.10",
-    "ts-jest": "^29.1.1",
-    "ts-loader": "^9.5.1",
-    "typescript": "^5.3.3",
-    "webpack": "^5.89.0",
-    "webpack-bundle-analyzer": "^4.10.1",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1",
-    "webpack-merge": "^5.10.0",
-    "webpack-pwa-manifest": "^4.3.0"
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   },
+  "workspaces": [
+    "frontend",
+    "backend",
+    "packages/*"
+  ],
   "engines": {
     "node": ">=18.0.0",
     "npm": ">=9.0.0"
   },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not op_mini all"
-  ],
   "lint-staged": {
-    "*.{ts,tsx}": [
+    "*.{js,jsx,ts,tsx}": [
       "eslint --fix",
       "prettier --write"
     ],
-    "*.{json,md,css}": [
+    "*.{json,css,scss,md}": [
       "prettier --write"
     ]
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gxjansen/llama-wool-farm.git"
+  },
+  "keywords": [
+    "llama",
+    "wool",
+    "farm",
+    "management",
+    "typescript",
+    "react",
+    "node",
+    "mongodb"
+  ],
+  "author": "Guido Jansen",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/gxjansen/llama-wool-farm/issues"
+  },
+  "homepage": "https://github.com/gxjansen/llama-wool-farm#readme"
 }
