@@ -1,50 +1,133 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  roots: ['<rootDir>/src', '<rootDir>/tests'],
-  testMatch: [
-    '**/__tests__/**/*.+(ts|tsx|js)',
-    '**/?(*.)+(spec|test).+(ts|tsx|js)',
-  ],
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
-  moduleNameMapper: {
+/** @type {import('jest').Config} */
+export default {
+  // Test environment
+  testEnvironment: 'node',
+  
+  // Setup files
+  setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
+  
+  // TypeScript support
+  preset: 'ts-jest/presets/default-esm',
+  extensionsToTreatAsEsm: ['.ts'],
+  
+  // Module resolution
+  moduleNameMapping: {
     '^@/(.*)$': '<rootDir>/src/$1',
-    '^@game/(.*)$': '<rootDir>/src/game/$1',
-    '^@scenes/(.*)$': '<rootDir>/src/scenes/$1',
-    '^@objects/(.*)$': '<rootDir>/src/objects/$1',
-    '^@systems/(.*)$': '<rootDir>/src/systems/$1',
-    '^@utils/(.*)$': '<rootDir>/src/utils/$1',
-    '^@types/(.*)$': '<rootDir>/src/types/$1',
-    '\\.(css|less|scss|sass)$': 'identity-obj-proxy',
-    '\\.(gif|ttf|eot|svg|png|jpg|jpeg|webp)$': '<rootDir>/tests/__mocks__/fileMock.js',
+    '^@shared/(.*)$': '<rootDir>/packages/shared/$1',
+    '^@frontend/(.*)$': '<rootDir>/frontend/src/$1',
+    '^@backend/(.*)$': '<rootDir>/backend/src/$1'
   },
-  setupFilesAfterEnv: [
-    '<rootDir>/tests/setup.ts',
-    'jest-canvas-mock',
+  
+  // Test file patterns
+  testMatch: [
+    '**/__tests__/**/*.(ts|tsx|js)',
+    '**/*.(test|spec).(ts|tsx|js)'
   ],
+  
+  // Coverage configuration
+  collectCoverage: true,
+  coverageDirectory: 'coverage',
+  coverageReporters: ['text', 'lcov', 'html', 'json'],
   collectCoverageFrom: [
     'src/**/*.{ts,tsx}',
-    '!src/**/*.d.ts',
-    '!src/**/index.ts',
-    '!src/types/**',
+    'frontend/src/**/*.{ts,tsx}',
+    'backend/src/**/*.{ts,tsx}',
+    'packages/shared/**/*.{ts,tsx}',
+    '\!**/*.d.ts',
+    '\!**/node_modules/**',
+    '\!**/dist/**',
+    '\!**/coverage/**',
+    '\!**/*.config.*',
+    '\!**/*.setup.*'
   ],
+  
+  // Coverage thresholds
   coverageThreshold: {
     global: {
-      branches: 70,
-      functions: 70,
-      lines: 70,
-      statements: 70,
-    },
+      branches: 80,
+      functions: 80,
+      lines: 80,
+      statements: 80
+    }
   },
-  coverageReporters: ['text', 'lcov', 'html'],
-  globals: {
-    'ts-jest': {
+  
+  // Transform configuration
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest', {
+      useESM: true,
       tsconfig: {
-        jsx: 'react',
+        target: 'ES2022',
+        module: 'ESNext',
+        moduleResolution: 'node',
         esModuleInterop: true,
-      },
-    },
+        allowSyntheticDefaultImports: true
+      }
+    }]
   },
+  
+  // Module file extensions
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json'],
+  
+  // Test timeout
+  testTimeout: 10000,
+  
+  // Ignore patterns
+  testPathIgnorePatterns: [
+    '<rootDir>/node_modules/',
+    '<rootDir>/dist/',
+    '<rootDir>/build/',
+    '<rootDir>/coverage/'
+  ],
+  
+  // Watch mode settings
+  watchPathIgnorePatterns: [
+    '<rootDir>/node_modules/',
+    '<rootDir>/dist/',
+    '<rootDir>/build/',
+    '<rootDir>/coverage/'
+  ],
+  
+  // Performance monitoring
+  maxWorkers: '50%',
+  
+  // Global setup and teardown
+  globalSetup: '<rootDir>/jest.global-setup.js',
+  globalTeardown: '<rootDir>/jest.global-teardown.js',
+  
+  // Verbose output
+  verbose: true,
+  
+  // Silent mode for CI
+  silent: process.env.CI === 'true',
+  
+  // Projects for monorepo support
+  projects: [
+    {
+      displayName: 'frontend',
+      testMatch: ['<rootDir>/frontend/**/*.test.(ts|tsx|js|jsx)'],
+      testEnvironment: 'jsdom',
+      setupFilesAfterEnv: ['<rootDir>/frontend/jest.setup.js']
+    },
+    {
+      displayName: 'backend',
+      testMatch: ['<rootDir>/backend/**/*.test.(ts|js)'],
+      testEnvironment: 'node',
+      setupFilesAfterEnv: ['<rootDir>/backend/jest.setup.js']
+    },
+    {
+      displayName: 'shared',
+      testMatch: ['<rootDir>/packages/shared/**/*.test.(ts|js)'],
+      testEnvironment: 'node'
+    }
+  ],
+  
+  // Clear mocks between tests
+  clearMocks: true,
+  
+  // Restore mocks after each test
+  restoreMocks: true,
+  
+  // Reset modules between tests
+  resetModules: true
 };
+EOF < /dev/null
