@@ -1,17 +1,36 @@
+const production = process.env.NODE_ENV === 'production';
+
 module.exports = {
   plugins: {
+    'postcss-import': {},
+    'postcss-preset-env': {
+      stage: 3,
+      features: {
+        'nesting-rules': true,
+        'custom-properties': true,
+        'custom-media-queries': true,
+      },
+      autoprefixer: {
+        flexbox: 'no-2009',
+      },
+    },
     'autoprefixer': {},
-    'cssnano': {
-      preset: ['default', {
-        discardComments: {
-          removeAll: true,
-        },
-        normalizeWhitespace: true,
-        colormin: true,
-        convertValues: true,
-        reduceIdents: true,
-        mergeRules: true,
-      }]
-    }
+    ...(production && {
+      'cssnano': {
+        preset: ['default', {
+          discardComments: {
+            removeAll: true,
+          },
+          normalizeWhitespace: true,
+          colormin: true,
+          convertValues: true,
+          reduceIdents: true,
+          mergeRules: true,
+          minifyFontValues: true,
+          minifyGradients: true,
+          minifySelectors: true,
+        }]
+      }
+    })
   }
 };
